+++ pages/schedule/[year]/[month]/[date].tsx
@@ -6,8 +6,7 @@ import { CalendarType } from '../../../../components/Header/CalendarTypeDropDown
 import Sidebar from '../../../../components/Sidebar';
 import { useDateContext } from '../../../../contexts/DateContext';
 import { useSidebarContext } from '../../../../contexts/SidebarContext';
-
-import styles from './schedule.module.scss';
+import styles from '../../../schedule.module.scss';
 
 export default function SchedulePage() {
     const { yearNow, monthNow, dateNow } = useDateContext();
